fix mentor comments 1.add requirements.txt 2.separate functions from main to functions.py 3.startup_loader() for load data from files
