+++ requirements.txt
@@ -1,0 +1,3 @@
+pathlib~=1.0.1
+tabulate~=0.9.0
+pyowm~=3.3.0
